+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `create_or_replace_data_view` handles filenames with spaces (either escaped or
+  quoted).
 
 ## [1.1.0] - 2021-06-14
 ### Changed