+++ CHANGELOG.md
@@ -5,10 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `create_or_replace_data_view` handles filenames with spaces (either escaped or
+  quoted).
+
 ### Added
 - `%ntbl push datasets [PATH]` command
 - `%ntbl pull datasets [PATH]` command
 - Logging to `/var/log/noteable_magics.log` or locally to `/tmp/noteable_magics.log`
+
 
 ## [1.1.0] - 2021-06-14
 ### Changed