+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.2.0] - 2021-07-19
 ### Changed
 - `create_or_replace_data_view` handles filenames with spaces (either escaped or
   quoted).
@@ -14,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `%ntbl pull datasets [PATH]` command
 - Logging to `/var/log/noteable_magics.log` or locally to `/tmp/noteable_magics.log`
 
+### Fixed
+- Use `httpx` instead of `requests` for better http streaming support
 
 ## [1.1.0] - 2021-06-14
 ### Changed